+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 * Fixed errors in console when offline player's wolves killed zombies
 * Update notify permission for admins with villagedefense.admin.* added
 * Now doors amount in setup menu will be normal (was 2x more because 1 door is 2 blocks)
+* Fixed bug where power ups were staying in game even if we left it, now power ups will remove after 40 seconds after spawn
+* Added /vd top <stat> command to show top 10 users!
 
 ### 3.6.1 Hotfix (04.05.2018)
 * Fixed /vda command not working on 1.8